Silva2016 lastree
Former-commit-id: 8144e9c86c995a901cce440c1391e74f2428239f
Former-commit-id: 99bc93f76099c2927070c1401ea8760b8cf69e0d
Former-commit-id: 5c6e3235c53c7c456a64044f0ed5bafd8ab6d011 [formerly 2ab5f6e5d68156aac0498db834b4e7650bc1025f]
Former-commit-id: 2424052a884a8f2082a4ee26131824294edb0130
Former-commit-id: d959a2a4cd368fc44a383352b5f6c0ffa9fe70b3
Former-commit-id: 22d1f44b4b46318ee61e4d3a18e6cca6418a33a2

+++ src/morphology.cpp
@@ -38,7 +38,7 @@ using namespace Rcpp;
 // [[Rcpp::export]]
 NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVector Z, double resolution, bool displaybar = false)
 {
-  int n = X.length();
+  long n = X.length();
   double half_res = resolution / 2;
 
   NumericVector Z_temp = clone(Z);
@@ -49,7 +49,7 @@ NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVect
   Progress p(2*n, displaybar);
 
   // Dilate
-  for (int i = 0 ; i < n ; i++)
+  for (long i = 0 ; i < n ; i++)
   {
     if (Progress::check_abort() )
       return Z_out;
@@ -61,7 +61,7 @@ NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVect
 
     double min_pt(std::numeric_limits<double>::max());
 
-    for(int j = 0 ; j < pts.size() ; j++)
+    for(long j = 0 ; j < pts.size() ; j++)
     {
       double z = Z_temp[pts[j]->id];
 
@@ -75,7 +75,7 @@ NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVect
   Z_temp = clone(Z_out);
 
   // erode
-  for (int i = 0 ; i < n ; i++)
+  for (long i = 0 ; i < n ; i++)
   {
     if (Progress::check_abort() )
       return Z_out;
@@ -87,7 +87,7 @@ NumericVector MorphologicalOpening(NumericVector X, NumericVector Y, NumericVect
 
     double max_pt(std::numeric_limits<double>::min());
 
-    for(int j = 0 ; j < pts.size() ; j++)
+    for(long j = 0 ; j < pts.size() ; j++)
     {
       double z = Z_temp[pts[j]->id];
